Move atomic file IO into core fs
Constraint: Atomic JSON/file persistence is shared infrastructure for storage modules, not a generic utilities bucket concern.
Rejected: Leave atomic-file in src/utils | storage owners would keep depending on an unclear catch-all directory.
Confidence: high
Scope-risk: moderate
Directive: Use src/core/fs for shared filesystem persistence primitives.
Tested: npm test -- --runTestsByPath tests/session/state-persistence.test.ts tests/storage-state/origin-scoped-restore.test.ts tests/storage-state/storage-state-manager.test.ts tests/tools/session-snapshot.test.ts tests/core/task-ledger/store.test.ts tests/core/task-run/storage.test.ts tests/core/trace/storage.test.ts tests/core/skill/storage.test.ts tests/core/skill-memory/store.test.ts tests/core/crawl/job-store.test.ts --runInBand; npm test -- --runTestsByPath tests/storage-state/storage-state-manager.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check
Not-tested: GitHub PR checks not yet run
