+++ src/core/crawl/job-store.ts
@@ -25,7 +25,7 @@ import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
 
-import { acquireLock } from '../../utils/atomic-file';
+import { acquireLock } from '../fs/atomic-file';
 import { normalizeUrl } from './utils';
 import { redactValue } from '../trace/redactor';
 import type { CrawlCacheMetadata, CrawlCacheMode, CrawlCacheScope } from './content-cache';
