+++ src/core/fs/atomic-file.ts
