+++ src/core/task-ledger/store.ts
@@ -29,7 +29,7 @@ import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
 
-import { acquireLock, writeFileAtomicSafe } from '../../utils/atomic-file';
+import { acquireLock, writeFileAtomicSafe } from '../fs/atomic-file';
 import type {
   TaskEvent,
   TaskKind,
